starting refactor working on file structure
